Move test classes to proper package (#47)
* Move test classes to proper package
* Package name changes.
